requirements.txt
Authlib==0.15.5
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
Flask==2.0.2
Flask-Authlib-Client==0.0.1
Flask-Login==0.4.1
flask-ngrok==0.0.25
Flask-WTF==1.1.1
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.0.1
packaging==21.3
phonenumbers==8.13.8
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==19.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-jose==3.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
